+++ src/baldr/attributes_controller.cc
@@ -85,6 +85,7 @@ const std::unordered_map<std::string, bool> AttributesController::kDefaultAttrib
     {kEdgeCountryCrossing, true},
     {kEdgeForward, true},
     {kEdgeLevels, true},
+    {kEdgeTrafficSignal, true},
 
     // Node keys
     {kIncidents, false},